+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Automatically open the current file if Rustlings is running in a VS Code terminal
 - Automatically open the current file with `$EDITOR` in a new pane if Rustlings is running in [Zellij](https://zellij.dev)
+- New argument `--no-editor` to disable automatic opening of the current file in VS Code or Zellij
 - New argument `--edit-cmd` to communicate with an editor running in a different process to open the current exercise
 - Show the file link of the current exercise when running `rustlings hint` and `rustlings reset`